change folder to elements and file names for base classes
