+++ EasyReflectometry/sample/assemblies/base_assembly.py
