+++ tests/sample/assemblies/test_surfactant_layer.py
@@ -7,8 +7,8 @@ Tests for Item class module
 import unittest
 
 from EasyReflectometry.sample.assemblies.surfactant_layer import SurfactantLayer
-from EasyReflectometry.sample.elementals.layers.layer import Layer
-from EasyReflectometry.sample.elementals.materials.material import Material
+from EasyReflectometry.sample.elements.layers.layer import Layer
+from EasyReflectometry.sample.elements.materials.material import Material
 
 
 class TestSurfactantLayer(unittest.TestCase):
