+++ tests/sample/elements/materials/test_material_density.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 from numpy.testing import assert_almost_equal
 
-from EasyReflectometry.sample.elementals.materials.material_density import MaterialDensity
+from EasyReflectometry.sample.elements.materials.material_density import MaterialDensity
 
 
 class TestMaterialDensity(unittest.TestCase):
